pyproject.toml
[tool.pytest.init_options]
testpaths = [
    "tests"
]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
    "services"
]

[tool.poetry]
name = "testss"
version = "0.1.0"
description = ""
authors = ["bevzd <dbevzpr@mail.ru>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.auth.dependencies]
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.11"
aiosignal = "1.3.2"
alembic = "1.14.0"
annotated-types = "0.7.0"
anyio = "4.7.0"
attrs = "24.3.0"
click = "8.1.8"
fastapi = "0.115.6"
fastapi-pagination = "0.12.34"
fastapi-sqlalchemy-toolkit = "0.7.11"
frozenlist = "1.5.0"
greenlet = "3.1.1"
grpcio = "1.68.1"
h11 = "0.14.0"
httptools = "0.6.4"
idna = "3.10"
mako = "1.3.8"
markupsafe = "3.0.2"
multidict = "6.1.0"
packaging = "24.2"
propcache = "0.2.1"
protobuf = "4.25.5"
pydantic = "2.10.4"
pydantic-settings = "2.7.0"
pydantic-core = "2.27.2"
python-dotenv = "1.0.1"
python-multipart = "0.0.20"
pyyaml = "6.0.2"
sniffio = "1.3.1"
sqlalchemy = "2.0.36"
sqlmodel = "0.0.22"
starlette = "0.41.3"
typing-extensions = "4.12.2"
uvicorn = "0.34.0"
uvloop = "0.21.0"
watchfiles = "1.0.3"
websockets = "14.1"
yarl = "1.18.3"
ydb = "3.18.12"
ydb-dbapi = "0.1.6"
ydb-sqlalchemy = "0.1.5"
fastapi-libkit = "^0.1.1"
passlib = "^1.7.4"
aiobotocore = "^2.16.1"
pyjwt = "^2.10.1"
httpx-oauth = "^0.16.1"


[tool.poetry.group.api.dependencies]
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.11"
aiosignal = "1.3.2"
alembic = "1.14.0"
annotated-types = "0.7.0"
anyio = "4.7.0"
arrow = "1.3.0"
attrs = "24.3.0"
backoff = "2.2.1"
certifi = "2024.12.14"
charset-normalizer = "3.4.0"
click = "8.1.8"
colorama = "0.4.6"
fastapi = "0.115.6"
fqdn = "1.5.1"
frozenlist = "1.5.0"
graphql-core = "3.2.5"
greenlet = "3.1.1"
grpcio = "1.68.1"
h11 = "0.14.0"
harfile = "0.3.0"
httpcore = "1.0.7"
httpx = "0.28.1"
hypothesis = "6.123.0"
hypothesis-graphql = "0.11.1"
hypothesis-jsonschema = "0.23.1"
idna = "3.10"
iniconfig = "2.0.0"
isoduration = "20.11.0"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
junit-xml = "1.9"
mako = "1.3.8"
markupsafe = "3.0.2"
multidict = "6.1.0"
packaging = "24.2"
pluggy = "1.5.0"
propcache = "0.2.1"
protobuf = "4.25.5"
pydantic = "2.10.4"
pydantic-settings = "2.7.0"
pydantic-core = "2.27.2"
pyrate-limiter = "3.7.0"
pytest = "8.3.4"
pytest-subtests = "0.14.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
referencing = "0.35.1"
requests = "2.32.3"
rfc3339-validator = "0.1.4"
rfc3987 = "1.3.8"
rpds-py = "0.22.3"
schemathesis = "3.39.2"
six = "1.17.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
sqlalchemy = "2.0.36"
starlette = "0.41.3"
starlette-testclient = "0.4.1"
tomli = "2.2.1"
tomli-w = "1.1.0"
types-python-dateutil = "2.9.0.20241206"
typing-extensions = "4.12.2"
uri-template = "1.3.0"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
webcolors = "24.11.1"
werkzeug = "3.1.3"
yarl = "1.18.3"
ydb = "3.18.12"
ydb-dbapi = "0.1.6"
ydb-sqlalchemy = "0.1.5"


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
